Change Interface of  NDArray & TBlob for DLPack Compatible (#6345)
* Change Interface of NDArray & TBlob for DLPack Compatible

Fix for cudnn operator

Fix cpp tests

* Update nnvm

* Fix for MKL mem

* Fix for windows macro

* Bump up version number to 0.10.1

* Update NDArray Save&Load

* trigger update

* Add test for legacy data load

* Use LegacyTShapeLoad

* trigger update

* Update tensor_blob.h

+++ src/io/image_io.cc
@@ -26,7 +26,7 @@ namespace io {
 // http://www.64lines.com/jpeg-width-height
 // Gets the JPEG size from the array of data passed to the function,
 // file reference: http://www.obrador.com/essentialjpeg/headerinfo.htm
-bool get_jpeg_size(const uint8_t* data, uint32_t data_size, uint32_t *width, uint32_t *height) {
+bool get_jpeg_size(const uint8_t* data, uint32_t data_size, int64_t *width, int64_t *height) {
   // Check for valid JPEG image
   uint32_t i = 0;  // Keeps track of the position within the file
   if (data[i] == 0xFF && data[i+1] == 0xD8 && data[i+2] == 0xFF && data[i+3] == 0xE0) {
@@ -63,7 +63,7 @@ bool get_jpeg_size(const uint8_t* data, uint32_t data_size, uint32_t *width, uin
   }
 }
 
-bool get_png_size(const uint8_t* data, uint32_t data_size, uint32_t *width, uint32_t *height) {
+bool get_png_size(const uint8_t* data, uint32_t data_size, int64_t *width, int64_t *height) {
   if (data[0] == 0x89 && data[1] == 0x50 && data[2] ==0x4E && data[3] == 0x47) {
     uint8_t const* p = data + 16;
     *width = ((p[0]*256 + p[1])*256 + p[2])*256 + p[3];
